building crypto. This is just prior to running coin gecko api script

+++ go/model_request_auction_bid.go
@@ -11,9 +11,10 @@
 package openapi
 
 type RequestAuctionBid struct {
+
 	Item string `json:"item,omitempty"`
 
-	Bid int32 `json:"bid,omitempty"`
+	Bid float32 `json:"bid,omitempty"`
 }
 
 // AssertRequestAuctionBidRequired checks if the required fields are not zero-ed
